@@ -14,7 +14,7 @@ dependencies:
   hive: ^2.2.3
   http: ^1.1.1
   intl: ^0.19.0
-  onesignal_flutter: ^5.0.4
+  onesignal_flutter: ^5.2.2
   path: ^1.8.3
   path_provider: ^2.1.0
   provider: ^6.1.2
@@ -31,10 +31,11 @@ dependencies:
   image: ^4.3.0
   geocoding: ^3.0.0
   geolocator: ^13.0.2
-  device_info_plus: ^11.1.1
+
   permission_handler: ^11.3.1
   sembast: ^3.8.0
   signature: ^5.5.0
+  async: ^2.12.0
 
 
 dev_dependencies:
