@@ -17,28 +17,25 @@ class TextResultCard extends StatelessWidget {
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: <Widget>[
         Container(
-          margin: const EdgeInsets.all(
-              4), // Replaced ScreenUtil with constant padding
-          width: MediaQuery.of(context).size.width /
-              4, // Responsive width based on screen size
+          margin: const EdgeInsets.all(4),
+          width: MediaQuery.of(context).size.width / 4,
           child: Text(
             title,
             style: TextStyle(
               color: colorBlack,
-              fontSize: 15, // Fixed font size
+              fontSize: 16,
             ),
           ),
         ),
         Expanded(
           child: Container(
-            margin: const EdgeInsets.only(
-                right: 4), // Replaced ScreenUtil with constant padding
+            margin: const EdgeInsets.only(right: 4),
             child: Text(
               value,
               textAlign: TextAlign.right,
               style: const TextStyle(
                 color: Colors.black,
-                fontSize: 15, // Fixed font size
+                fontSize: 16,
               ),
             ),
           ),
