@@ -11,14 +11,14 @@ class DashboardPPTabController extends GetxController
 
   void initController(TickerProvider vsync) {
     controller =
-        TabController(vsync: vsync, length: 2, initialIndex: initialIndex);
+        TabController(vsync: vsync, length: 4, initialIndex: initialIndex);
   }
 
   @override
   void onInit() {
     super.onInit();
     controller =
-        TabController(vsync: this, length: 2, initialIndex: initialIndex);
+        TabController(vsync: this, length: 4, initialIndex: initialIndex);
   }
 
   @override
@@ -27,4 +27,3 @@ class DashboardPPTabController extends GetxController
     super.onClose();
   }
 }
-// d51e7b74-eebc-48e9-8af6-a2d1cbd58e33
