@@ -6,6 +6,7 @@ import 'package:intl/intl.dart';
 import 'package:noo_sms/assets/constant/api_constant.dart';
 import 'package:noo_sms/models/approval_status.dart';
 import 'package:noo_sms/models/noo_approval.dart';
+
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart' as http;
 import 'package:signature/signature.dart';
@@ -40,6 +41,7 @@ class ApprovalController extends GetxController {
   final RxInt idUser = 0.obs;
   final RxInt editApproval = 0.obs;
   final RxString signatureApprovalFromServer = ''.obs;
+  final Map<int, Map<String, dynamic>> _approvalControllers = {};
 
   @override
   void onInit() {
@@ -49,26 +51,85 @@ class ApprovalController extends GetxController {
         'Basic ${base64Encode(utf8.encode('$usernameAuth:$passwordAuth'))}';
   }
 
+  void initializeApprovalControllers(int id) {
+    if (!_approvalControllers.containsKey(id)) {
+      _approvalControllers[id] = {
+        'remark': TextEditingController(),
+        'paymentTerm': TextEditingController(),
+        'creditLimit': TextEditingController(),
+        'signature': SignatureController(
+          penStrokeWidth: 1,
+          penColor: Colors.black,
+          exportBackgroundColor: Colors.white,
+        ),
+        'selectedPaymentTerm': '',
+        'creditLimitError': '',
+      };
+      update(['approval-$id']);
+    }
+  }
+
   Future<void> initializeData() async {
     await getSharedPrefs();
     await fetchApprovals();
   }
 
+  TextEditingController getRemarkController(int id) {
+    return _approvalControllers[id]?['remark'] ?? TextEditingController();
+  }
+
+  TextEditingController getCreditLimitController(int id) {
+    return _approvalControllers[id]?['creditLimit'] ?? TextEditingController();
+  }
+
+  SignatureController getSignatureController(int id) {
+    return _approvalControllers[id]?['signature'] ??
+        SignatureController(
+          penStrokeWidth: 1,
+          penColor: Colors.black,
+          exportBackgroundColor: Colors.white,
+        );
+  }
+
+  void setSelectedPaymentTerm(int id, String value) {
+    if (!_approvalControllers.containsKey(id)) {
+      initializeApprovalControllers(id);
+    }
+    _approvalControllers[id]!['selectedPaymentTerm'] = value; // Store as String
+    update(['payment-terms-$id']);
+  }
+
+  String getSelectedPaymentTerm(int id) {
+    if (!_approvalControllers.containsKey(id)) {
+      initializeApprovalControllers(id);
+    }
+    return _approvalControllers[id]!['selectedPaymentTerm'] as String? ?? '';
+  }
+
+  void setCreditLimitError(int id, String value) {
+    if (!_approvalControllers.containsKey(id)) {
+      initializeApprovalControllers(id);
+    }
+    _approvalControllers[id]!['creditLimitError'] = value;
+    update(['credit-limit-error-$id']);
+  }
+
+  String getCreditLimitError(int id) {
+    return (_approvalControllers[id]?['creditLimitError'] as String?) ?? '';
+  }
+
   Future<void> getSharedPrefs() async {
     final prefs = await SharedPreferences.getInstance();
-    idUser.value = prefs.getInt("iduser") ?? 0;
+    idUser.value = prefs.getInt("userid") ?? 0;
     editApproval.value = prefs.getInt("editApproval") ?? 0;
   }
 
   Future<void> fetchApprovals() async {
     isLoading.value = true;
     final prefs = await SharedPreferences.getInstance();
-    final userId = prefs.getInt("iduser")?.toString();
-    final String? role = prefs.getString("Role");
+    final userId = prefs.getInt("userid")?.toString();
 
-    final url = role == "1"
-        ? '$baseURLDevelopment/FindApproval/$userId?page=${page.value}'
-        : '$baseURLDevelopment/ViewAllCust?page=${page.value}';
+    final url = '${baseURLDevelopment}FindApproval/$userId?page=${page.value}';
 
     final response = await makeApiCall(url);
 
@@ -80,58 +141,162 @@ class ApprovalController extends GetxController {
   }
 
   Future<void> fetchApprovalDetail(int id) async {
-    isLoading.value = true;
-    final response = await makeApiCall('$baseURLDevelopment/NOOCustTables/$id');
-    debugPrint(response.statusCode.toString());
-    if (response.statusCode == 200) {
-      final data = json.decode(response.body);
-      currentApproval.value = ApprovalModel.fromJson(data);
-      companyAddress.value = Address.fromJson(data['CompanyAddresses']);
-      deliveryAddress.value = Address.fromJson(data['DeliveryAddresses']);
-      taxAddress.value = Address.fromJson(data['TaxAddresses']);
+    try {
+      isLoading.value = true;
+      initializeApprovalControllers(id);
 
-      await Future.wait([
-        fetchCreditLimitRange(currentApproval.value.segment ?? ''),
-        fetchPaymentTerms(currentApproval.value.segment ?? ''),
-        fetchApprovalStatuses(id),
-      ]);
-      selectedPaymentTerm.value =
-          currentApproval.value.paymentTerm ?? paymentTerms.first;
-      paymentTermController.text = selectedPaymentTerm.value;
+      final response =
+          await makeApiCall('${baseURLDevelopment}NOOCustTables/$id');
+
+      if (response.statusCode == 200) {
+        final data = json.decode(response.body);
+        currentApproval.value = ApprovalModel.fromJson(data);
+        companyAddress.value = Address.fromJson(data['CompanyAddresses']);
+        deliveryAddress.value = Address.fromJson(data['DeliveryAddresses']);
+        taxAddress.value = Address.fromJson(data['TaxAddresses']);
+
+        // Load all required data
+        await Future.wait([
+          fetchCreditLimitRange(currentApproval.value.segment ?? ''),
+          fetchPaymentTerms(currentApproval.value.segment ?? ''),
+          fetchApprovalStatuses(id),
+        ]);
+
+        // Set initial values after data is loaded
+        getCreditLimitController(id).text =
+            currentApproval.value.creditLimit?.toString() ?? '';
+
+        final currentPaymentTerm = currentApproval.value.paymentTerm;
+        if (currentPaymentTerm != null && currentPaymentTerm.isNotEmpty) {
+          setSelectedPaymentTerm(id, currentPaymentTerm);
+        } else if (paymentTerms.isNotEmpty) {
+          setSelectedPaymentTerm(id, paymentTerms.first);
+        }
+
+        update(['approval-$id']);
+      }
+    } catch (e) {
+      debugPrint('Error in fetchApprovalDetail: $e');
+    } finally {
+      isLoading.value = false;
     }
   }
 
-  Future<void> processApproval(int id, String approvedSignature, String remark,
+  String formatNumberForDisplay(String value) {
+    try {
+      final cleanValue = value.replaceAll('.', '');
+
+      final number = double.parse(cleanValue);
+
+      final formatter = NumberFormat('#,###', 'id_ID');
+      return formatter.format(number).replaceAll(',', '.');
+    } catch (e) {
+      return '0';
+    }
+  }
+
+  String formatForApi(String value) {
+    try {
+      // Remove all dots and convert to double
+      final cleanValue = value.replaceAll('.', '');
+      return double.parse(cleanValue).toString();
+    } catch (e) {
+      return '0';
+    }
+  }
+
+  Future<void> processApproval(BuildContext context, int id, String remark,
       bool isCreditLimitValid) async {
     try {
       if (!isCreditLimitValid) {
-        // showErrorAlert(
-        //     'Credit limit must be between ${minCreditLimit.value} and ${maxCreditLimit.value}');
         return;
       }
 
-      final url = '$baseURLDevelopment/ApprovalModel?'
+      final signatureController = getSignatureController(id);
+      final signature = await signatureController.toPngBytes();
+
+      if (signature == null) {
+        Get.snackbar('Error', 'Please provide signature');
+        return;
+      }
+
+      // Show loading indicator
+      Get.dialog(
+        const Center(
+          child: CircularProgressIndicator(),
+        ),
+        barrierDismissible: false,
+      );
+
+      await uploadSignature(signature);
+
+      final apiCreditLimit = formatForApi(getCreditLimitController(id).text);
+
+      final url = '${baseURLDevelopment}Approval?'
           'id=$id&'
           'value=1&'
           'approveBy=${idUser.value}&'
-          'ApprovedSignature=$approvedSignature&'
+          'ApprovedSignature=${signatureApprovalFromServer.value}&'
           'Remark=$remark&'
-          'paymentId=${paymentTermController.text}&'
-          'creditLimit=${creditLimitController.text.replaceAll(",", "")}';
+          'paymentId=${getSelectedPaymentTerm(id)}&'
+          'creditLimit=$apiCreditLimit';
 
       final response = await makeApiCall(url, method: 'POST');
 
       if (response.statusCode == 200) {
-        Get.offAllNamed('/dashboard');
-        showSuccessMessage('ApprovalModel processed successfully');
+        _disposeApprovalControllers(id);
+
+        Get.back();
+
+        Get.snackbar(
+          'Success',
+          'Approval processed successfully',
+          snackPosition: SnackPosition.BOTTOM,
+          backgroundColor: Colors.green,
+          colorText: Colors.white,
+          duration: const Duration(seconds: 2),
+        );
+
+        await Future.delayed(const Duration(milliseconds: 500));
+        Get.toNamed('/noo_approve');
+      } else {
+        Get.back();
+        handleError('Failed to process approval',
+            'Server returned ${response.statusCode}');
       }
     } catch (e) {
-      handleError('Error processing approval', e);
+      if (Get.isDialogOpen ?? false) {
+        Get.back();
+      }
+      handleError('Error processing approval', e.toString());
     }
   }
 
+// Update show success message to return Future
+  Future<void> showSuccessMessage(String message) async {
+    Get.snackbar(
+      'Success',
+      message,
+      snackPosition: SnackPosition.BOTTOM,
+      backgroundColor: Colors.green,
+      colorText: Colors.white,
+      duration: const Duration(seconds: 2),
+    );
+  }
+
+  void handleError(String message, String errorDetail) {
+    Get.snackbar(
+      'Error',
+      '$message\n$errorDetail',
+      snackPosition: SnackPosition.BOTTOM,
+      backgroundColor: Colors.red,
+      colorText: Colors.white,
+      duration: const Duration(seconds: 3),
+    );
+  }
+
   Future<void> processReject(int id, String remark) async {
-    final url = '$baseURLDevelopment/ApprovalModel/$id?'
+    final url = '${baseURLDevelopment}ApprovalModel/$id?'
         'value=0&'
         'approveBy=${idUser.value}&'
         'ApprovedSignature=reject&'
@@ -140,7 +305,7 @@ class ApprovalController extends GetxController {
     final response = await makeApiCall(url, method: 'POST');
 
     if (response.statusCode == 200) {
-      Get.offAllNamed('/dashboard');
+      // Get.offAllNamed('/dashboard');
       showSuccessMessage('Rejection processed successfully');
     }
   }
@@ -152,7 +317,7 @@ class ApprovalController extends GetxController {
         "SIGNATUREAPPROVAL_${DateFormat("ddMMyyyy_hhmmss").format(DateTime.now())}_${username}_.jpg";
 
     var request =
-        http.MultipartRequest('POST', Uri.parse('$baseURLDevelopment/Upload'))
+        http.MultipartRequest('POST', Uri.parse('${baseURLDevelopment}Upload'))
           ..files.add(http.MultipartFile.fromBytes(
             'file',
             signature,
@@ -167,7 +332,7 @@ class ApprovalController extends GetxController {
   Future<void> fetchCreditLimitRange(String segment) async {
     final adjustedSegment = segment != "Hotels" ? "All" : segment;
     final response = await makeApiCall(
-        '$baseURLDevelopment/CreditLimits/BySegment/$adjustedSegment');
+        '${baseURLDevelopment}CreditLimits/BySegment/$adjustedSegment');
 
     if (response.statusCode == 200) {
       final data = json.decode(response.body);
@@ -176,19 +341,21 @@ class ApprovalController extends GetxController {
     }
   }
 
-  void validateCreditLimit(String value) {
+  void validateCreditLimit(int id, String value) {
     try {
-      final creditLimit = double.tryParse(value.replaceAll(",", "")) ?? 0;
+      final cleanValue = value.replaceAll('.', '');
+      final creditLimit = double.parse(cleanValue);
       if (creditLimit < minCreditLimit.value ||
           creditLimit > maxCreditLimit.value) {
-        creditLimitError.value =
+        setCreditLimitError(
+            id,
             "Credit limit must be between ${formatCurrency(minCreditLimit.value)} "
-            "and ${formatCurrency(maxCreditLimit.value)}";
+            "and ${formatCurrency(maxCreditLimit.value)}");
       } else {
-        creditLimitError.value = '';
+        setCreditLimitError(id, '');
       }
     } catch (e) {
-      creditLimitError.value = 'Invalid number format';
+      setCreditLimitError(id, 'Invalid number format');
     }
   }
 
@@ -222,7 +389,7 @@ class ApprovalController extends GetxController {
   Future<void> fetchApprovalStatuses(int id) async {
     try {
       final response = await http.get(
-        Uri.parse('$baseURLDevelopment/Approval/$id'),
+        Uri.parse('${baseURLDevelopment}Approval/$id'),
         headers: {'authorization': basicAuth},
       );
 
@@ -237,46 +404,37 @@ class ApprovalController extends GetxController {
   }
 
   Future<void> fetchPaymentTerms(String segment) async {
-    var url = "${baseURLDevelopment}PaymentTerms/BySegment/$segment";
-    String basicAuth = 'Basic ${base64Encode(utf8.encode('test:test456'))}';
-    final response = await http.get(Uri.parse(url),
-        headers: <String, String>{'authorization': basicAuth});
-    List<dynamic> jsonResponse = json.decode(response.body);
+    try {
+      var url = "${baseURLDevelopment}PaymentTerms/BySegment/$segment";
+      final response = await makeApiCall(url);
 
-    paymentTerms.value =
-        jsonResponse.map((item) => item['PaymTermId'].toString()).toList();
+      if (response.statusCode == 200) {
+        List<dynamic> jsonResponse = json.decode(response.body);
+        paymentTerms.value =
+            jsonResponse.map((item) => item['PaymTermId'].toString()).toList();
+        update(['payment-terms']);
+      }
+    } catch (e) {
+      debugPrint('Error fetching payment terms: $e');
+    }
   }
 
-  void handleError(String message, dynamic error) {
-    Get.snackbar(
-      'Error',
-      message,
-      snackPosition: SnackPosition.BOTTOM,
-      backgroundColor: Colors.red,
-      colorText: Colors.white,
-    );
+  void _disposeApprovalControllers(int id) {
+    final controllers = _approvalControllers[id];
+    if (controllers != null) {
+      controllers['remark'].dispose();
+      controllers['paymentTerm'].dispose();
+      controllers['creditLimit'].dispose();
+      controllers['signature'].dispose();
+      _approvalControllers.remove(id);
+    }
   }
-
-  void showSuccessMessage(String message) {
-    Get.snackbar(
-      'Success',
-      message,
-      snackPosition: SnackPosition.BOTTOM,
-      backgroundColor: Colors.green,
-      colorText: Colors.white,
-    );
-  }
-
-  // void showErrorAlert(String message) {
-  //   QuickAlert.show(
-  //     context: Get.context!,
-  //     type: QuickAlertType.error,
-  //     text: message,
-  //   );
-  // }
 
   @override
   void onClose() {
+    for (final id in _approvalControllers.keys.toList()) {
+      _disposeApprovalControllers(id);
+    }
     remarkController.dispose();
     paymentTermController.dispose();
     creditLimitController.dispose();
