@@ -191,8 +191,18 @@ class ApprovalModel {
             ? Address.fromJson(json['TaxAddresses'])
             : null,
         paymentTerm: json['PaymentTerm'],
-        creditLimit: json['CreditLimit']?.toDouble(),
+        creditLimit: _parseCreditLimit(json['CreditLimit']),
       );
+
+  static double? _parseCreditLimit(dynamic value) {
+    if (value == null) return null;
+    if (value is double) return value;
+    if (value is int) return value.toDouble();
+    if (value is String) {
+      return double.tryParse(value.replaceAll(',', ''));
+    }
+    return null;
+  }
 
   Map<String, dynamic> toJson() => {
         'id': id,
