@@ -38,7 +38,7 @@ class Promotion {
   dynamic businessUnit;
   String? price;
   String? totalAmount;
-  bool? status; // nullable if they can be null
+  bool? status;
   dynamic codeError;
   dynamic message;
   dynamic listId;
@@ -254,7 +254,10 @@ class Promotion {
       Uri.parse(url),
       headers: {'Authorization': token},
     );
+    debugPrint(token);
+    debugPrint(nomorPP);
 
+    debugPrint(response.body);
     if (response.statusCode == 200) {
       List jsonResponse = jsonDecode(response.body);
       return jsonResponse.map((e) => Promotion.fromJson(e)).toList();
@@ -310,6 +313,7 @@ class Promotion {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     String username = prefs.getString('username')!;
     String token = prefs.getString('token')!;
+    debugPrint("tzt$idProduct$idCustomer$token");
     String url =
         "$apiCons/api/SalesOrder?idProduct=$idProduct&idCustomer=$idCustomer&username=$username";
 
