@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:noo_sms/assets/global.dart';
 import 'package:noo_sms/controllers/dashboard/dashboard_ordertaking.dart';
-import 'package:noo_sms/view/promotion_program/order/create_order_taking.dart';
-import 'package:noo_sms/view/promotion_program/order/history_order_taking.dart';
+import 'package:noo_sms/view/sms/order/create_order_taking.dart';
+import 'package:noo_sms/view/sms/order/history_order_taking.dart';
 
 class DashboardOrderTaking extends StatefulWidget {
   final int initialIndex;
@@ -44,6 +44,10 @@ class _DashboardOrderTakingState extends State<DashboardOrderTaking>
           mainAxisSize: MainAxisSize.min,
           children: [
             TabBar(
+              labelStyle: const TextStyle(
+                fontWeight: FontWeight.w500,
+                fontSize: 16,
+              ),
               unselectedLabelColor: Colors.black,
               labelColor: colorAccent,
               controller: tabController.controller,
